+++ packages/product-feed-zbozi/tests/TestStandardFeedItem.php
@@ -259,6 +259,14 @@ class TestStandardFeedItem implements StandardFeedItemInterface
     }
 
     /**
+     * @return int
+     */
+    public function getMainCategoryId()
+    {
+        return 0;
+    }
+
+    /**
      * @param string $name
      * @return mixed
      */